Pipewire.CoreAPI.ThreadLoop

Description

The threaded loop enable using the binding from multiple thread.

Actions needs to be wrapped using withLock, and to let the loop run, for example to take a MVar, then use withUnlock.

TODO: investigate managing the thread loop manually, e.g. by waiting for `IO ()` action and pausing the loop with pw_main_loop_quit

Documentation

newtype ThreadLoop

Constructors

ThreadLoop (Ptr PwThreadLoopStruct)

thread_debug :: Bool

threadLog :: String -> IO ()

pw_thread_loop_new :: CString -> IO ThreadLoop

inline_c_ffi_6989586621679058756 :: Ptr CChar -> IO (Ptr PwThreadLoopStruct)

pw_thread_loop_lock :: ThreadLoop -> IO ()

inline_c_ffi_6989586621679058765 :: Ptr PwThreadLoopStruct -> IO ()

pw_thread_loop_unlock :: ThreadLoop -> IO ()

inline_c_ffi_6989586621679058774 :: Ptr PwThreadLoopStruct -> IO ()

withLock :: ThreadLoop -> IO a -> IO a

A lock must be held when using client implementation.

withUnlock :: ThreadLoop -> IO a -> IO a

Temporarly release a lock to let the loop run.

pw_thread_loop_get_loop :: ThreadLoop -> IO Loop

inline_c_ffi_6989586621679058785 :: Ptr PwThreadLoopStruct -> IO (Ptr PwLoopStruct)

pw_thread_loop_destroy :: ThreadLoop -> IO ()

inline_c_ffi_6989586621679058794 :: Ptr PwThreadLoopStruct -> IO ()

pw_thread_loop_wait :: ThreadLoop -> IO ()

inline_c_ffi_6989586621679058803 :: Ptr PwThreadLoopStruct -> IO ()

pw_thread_loop_start :: ThreadLoop -> IO ()

inline_c_ffi_6989586621679058812 :: Ptr PwThreadLoopStruct -> IO CInt

pw_thread_loop_stop :: ThreadLoop -> IO ()

inline_c_ffi_6989586621679058821 :: Ptr PwThreadLoopStruct -> IO ()

pw_thread_loop_signal :: ThreadLoop -> Bool -> IO ()

inline_c_ffi_6989586621679058833 :: Ptr PwThreadLoopStruct -> CBool -> IO ()

withThreadSignalsHandler :: ThreadLoop -> IO a -> IO a

Stop the loop on SIGINT or SIGKILL

inline_c_ffi_6989586621679058850 :: SignalHandlerRaw -> IO (FunPtr SignalHandlerRaw)

inline_c_ffi_6989586621679058863 :: Ptr PwThreadLoopStruct -> FunPtr (Ptr () -> CInt -> IO ()) -> FunPtr (Ptr () -> CInt -> IO ()) -> IO ()