Pipewire.Internal

Description

Helper functions and re-export

Documentation

peekCString :: CString -> IO Text

Base

data Word32

32-bit unsigned integer type

Foreign

data CInt

Haskell type representing the C int type. (The concrete types of Foreign.C.Types are platform-specific.)

type CString = Ptr CChar

A C string is a reference to an array of C characters terminated by NUL.

data Ptr a

A value of type Ptr a represents a pointer to an object, or an array of objects, which may be marshalled to or from Haskell values of type a.

The type a will often be an instance of class Storable which provides the marshalling operations. However this is not essential, and you can provide your own operations to access the pointer. For example you might write small foreign functions to get or set the fields of a C struct.

nullPtr :: Ptr a

The constant nullPtr contains a distinguished value of Ptr that is not associated with a valid memory location.

text

data Text

A space efficient, packed, unboxed Unicode text type.

withCString :: Text -> (CString -> IO a) -> IO a

Marshal a Text into a C string with a trailing NUL byte, encoded as UTF-8 in temporary storage.

The temporary storage is freed when the subcomputation terminates (either normally or via an exception), so the pointer to the temporary storage must not be used after this function returns.

Since: text-2.0.1