Pipewire

High level API

data PwInstance state

A pipewire instance

Constructors

PwInstance

Fields

stateVar :: MVar state

mainLoop :: PwMainLoop

core :: PwCore

registry :: PwRegistry

sync :: IORef SeqID

errorsVar :: MVar [CoreError]

data RegistryEvent

A registry event

Constructors

Added PwID Text SpaDict

Removed PwID

withInstance :: state -> (RegistryEvent -> state -> IO state) -> (PwInstance state -> IO a) -> IO a

Create a new PwInstance by providing an initial state and a registry update handler.

runInstance :: PwInstance state -> IO (Maybe (NonEmpty CoreError))

Run the main loop

quitInstance :: PwInstance state -> IO ()

Terminate the main loop, to be called from handlers.

data CoreError

A pipewire error

Constructors

CoreError

Fields

pwid :: PwID

code :: Int

message :: Text

syncState :: PwInstance state -> (Either (NonEmpty CoreError) state -> IO a) -> IO a

Ensure all the events have been processed and access the state.

syncState_ :: PwInstance state -> (state -> IO a) -> IO a

Like syncState but throwing an error if there was any pipewire error.

Mid level bracket API

withPipewire :: IO a -> IO a

withMainLoop :: (PwMainLoop -> IO a) -> IO a

withContext :: PwLoop -> (PwContext -> IO a) -> IO a

withCore :: PwContext -> (PwCore -> IO a) -> IO a

Protocol

newtype PwID

Constructors

PwID

Fields

getID :: Int

newtype PwVersion

Constructors

PwVersion

Fields

getVersion :: Int

newtype SeqID

Constructors

SeqID

Fields

getID :: Int

module Pipewire.Constants

module Pipewire.Enum

Core API

Initialization

pw_deinit :: IO ()

pw_init :: IO ()

Main Loop

data PwMainLoop

pw_main_loop_destroy :: PwMainLoop -> IO ()

pw_main_loop_get_loop :: PwMainLoop -> IO PwLoop

pw_main_loop_new :: IO PwMainLoop

pw_main_loop_quit :: PwMainLoop -> IO CInt

pw_main_loop_run :: PwMainLoop -> IO CInt

Context

data PwContext

pw_context_connect :: PwContext -> IO PwCore

pw_context_destroy :: PwContext -> IO ()

pw_context_new :: PwLoop -> IO PwContext

Core

type DoneHandler = PwID -> SeqID -> IO ()

type ErrorHandler = PwID -> SeqID -> Int -> Text -> IO ()

type InfoHandler = PwCoreInfo -> IO ()

data PwCore

data PwCoreEvents

data PwCoreInfo

data PwRegistry

pw_core_add_listener :: PwCore -> SpaHook -> PwCoreEvents -> IO ()

pw_core_disconnect :: PwCore -> IO ()

pw_core_get_registry :: PwCore -> IO PwRegistry

pw_core_sync :: PwCore -> PwID -> IO SeqID

pw_id_core :: PwID

with_pw_core_events :: InfoHandler -> DoneHandler -> ErrorHandler -> (PwCoreEvents -> IO b) -> IO b

Create a local pw_core_events structure

Link

type LinkState = Either Text PwLinkState

newtype PwLink

Constructors

PwLink

Fields

getProxy :: PwProxy

pw_link_create :: PwCore -> PwProperties -> IO PwLink

with_pw_link_events :: PwLink -> PwLinkEventInfoHandler -> IO a -> IO a

Loop

data PwLoop

Proxy

data PwProxy

pw_proxy_destroy :: PwProxy -> IO ()

with_pw_proxy_events :: PwProxy -> ProxyDestroyHandler -> ProxyRemovedHandler -> ProxyErrorHandler -> IO a -> IO a

Registry

type GlobalHandler = PwID -> Text -> PwVersion -> SpaDict -> IO ()

type GlobalRemoveHandler = PwID -> IO ()

pw_registry_add_listener :: PwRegistry -> SpaHook -> PwRegistryEvents -> IO ()

pw_registry_destroy :: PwRegistry -> PwID -> IO ()

with_pw_registry_events :: GlobalHandler -> GlobalRemoveHandler -> (PwRegistryEvents -> IO b) -> IO b

Create a local pw_registry_events structure

Utilities

Properties

data PwProperties

pw_properties_get :: PwProperties -> Text -> IO (Maybe Text)

pw_properties_new :: IO PwProperties

pw_properties_new_dict :: SpaDict -> IO PwProperties

pw_properties_set :: PwProperties -> Text -> Text -> IO ()

pw_properties_set_id :: PwProperties -> Text -> PwID -> IO ()

pw_properties_set_linger :: PwProperties -> IO ()

Keep the object after the client disconnect

SPA

Utilities

Dictionary

data SpaDict

spaDictLookup :: SpaDict -> Text -> IO (Maybe Text)

spaDictLookupInt :: SpaDict -> Text -> IO (Maybe Int)

spaDictRead :: SpaDict -> IO (Vector (Text, Text))

Read SpaDict keys/values

with_spa_dict :: (SpaDict -> IO a) -> IO a

Create a local spa_hook structure

Hooks

data SpaHook

with_spa_hook :: (SpaHook -> IO a) -> IO a

Create a local spa_hook structure

Helpers

getHeadersVersion :: IO Text

getLibraryVersion :: IO Text